(#10) Made ServiceLocator an interface and moved implementation into JerseyResourceLocator; also changed "generation" package to "generator" for consistency
